requirements.txt
acres==0.5.0
agentops==0.4.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apify_client==1.11.0
apify_shared==1.4.1
arrow==1.3.0
arxiv==2.2.0
arxiv2text==0.1.14
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
blinker==1.9.0
cachetools==5.5.2
camel-ai==0.2.68
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
cssselect==1.3.0
dappier==0.3.6
debugpy @ file:///D:/bld/debugpy_1744321277463/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
deepseek==1.0.0
Deprecated==1.2.18
distro==1.9.0
docstring-parser==0.15
docutils==0.21.2
duckduckgo_search==6.4.2
etelemetry==0.3.1
exa-py==1.14.10
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
faiss-cpu==1.11.0
fake-useragent==2.2.0
fastapi==0.115.13
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
firecrawl-py==1.17.0
free_proxy==1.1.3
frozenlist==1.7.0
fsspec==2025.5.1
geojson==2.5.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.166.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
googlemaps==4.10.0
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
imagesize==1.4.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711232/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jaraco.context==6.0.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
linkup-sdk==0.2.6
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mcp==1.9.4
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.5.0
narwhals==1.44.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.5
newspaper3k==0.2.8
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
openai==1.90.0
openai-agents==0.0.19
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathlib==1.0.1
pdfminer.six==20250506
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
playwright==1.52.0
primp==0.15.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.2
psutil==5.9.8
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
pyee==13.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyMuPDF==1.26.1
pyowm==3.3.0
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==307
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1749898448924/work
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==1.3.1
roman-numerals-py==3.1.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scholarly==1.7.11
scikit-learn==1.7.0
scipy==1.15.3
scrapegraph_py==1.12.0
selenium==4.33.0
sentence-transformers==4.1.0
sgmllib3k==1.0.0
simplejson==3.20.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sse-starlette==2.3.6
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.2
streamlit==1.46.0
sympy==1.14.0
tavily-python==0.5.4
tenacity==9.1.2
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado @ file:///D:/bld/tornado_1748003321906/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
traits==7.0.2
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
